Fix integer truncation in process start time to preserve sub-second precision (#115)
Fix integer truncation in process start time to preserve sub-second
precision

### Motivation:

This PR fixes a bug where the `process_start_time_seconds` metric was
being calculated using integer division, causing the sub-second
fractional component of the start time to be silently truncated. This
resulted in the reported metric being inaccurate by up to ~1 second.

### Modifications:

- Updated the `SystemMetricsMonitor.Data` struct to store
`startTimeSeconds` as a `Double` instead of an `Int` to preserve
fractional seconds.
- Linux: Converted `startTimeTicks` and `clockTicksPerSecond` to
`Double` prior to division in `SystemMetricsMonitor+Linux.swift` to
prevent integer truncation.
- Darwin: Included the `pbi_start_tvusec` (microseconds) field from
`kinfo_proc` to properly calculate the fractional component in
`SystemMetricsMonitor+Darwin.swift`.
- Updated `DarwinDataProviderTests` to compare the `Double` start time
metric against the full sub-second timestamp provided by `sysctl`.

### Result:

This aligns the process start time precision with how CPU time is
currently computed and with Prometheus metric conventions.

+++ Sources/SystemMetrics/SystemMetricsMonitor.swift
@@ -271,7 +271,7 @@ extension SystemMetricsMonitor {
         /// Resident memory size in bytes.
         package var residentMemoryBytes: Int
         /// Start time of the process since UNIX epoch in seconds.
-        package var startTimeSeconds: Int
+        package var startTimeSeconds: Double
         /// Total user and system CPU time spent in seconds.
         package var cpuSeconds: Double
         /// Maximum number of open file descriptors.
@@ -294,7 +294,7 @@ extension SystemMetricsMonitor {
         package init(
             virtualMemoryBytes: Int,
             residentMemoryBytes: Int,
-            startTimeSeconds: Int,
+            startTimeSeconds: Double,
             cpuSeconds: Double,
             maxFileDescriptors: Int,
             openFileDescriptors: Int,
